fix: zoom feature for images in docs and blog - rebuild structure
