+++ src/components/shared/content/content.jsx
@@ -1,5 +1,6 @@
 /* eslint-disable react/prop-types */
 import clsx from 'clsx';
+import Image from 'next/image';
 import { MDXRemote } from 'next-mdx-remote/rsc';
 import PropTypes from 'prop-types';
 import { Fragment } from 'react';
@@ -24,8 +25,7 @@ import getCodeProps from 'lib/rehype-code-props';
 
 import sharedMdxComponents from '../../../../content/docs/shared-content';
 import DocCta from '../doc-cta';
-
-import ImageBlock from './image-block'
+import ImageZoom from '../image-zoom'
 
 const sharedComponents = Object.keys(sharedMdxComponents).reduce((acc, key) => {
   acc[key] = () => IncludeBlock({ url: sharedMdxComponents[key] });
@@ -87,10 +87,20 @@ const getComponents = (withoutAnchorHeading, isReleaseNote, isPostgres) => ({
     );
   },
   img: (props) => {
-    const { className, title, ...rest } = props;
+    const { className, title, src, ...rest } = props;
 
     return (
-      <ImageBlock className={className} title={title} isReleaseNote={isReleaseNote} {...rest} />
+      <ImageZoom src={src}>
+        <Image
+          className={clsx(className, { 'no-border': title === 'no-border' })}
+          src={src}
+          width={isReleaseNote ? 762 : 796}
+          height={isReleaseNote ? 428 : 447}
+          style={{ width: '100%', height: '100%' }}
+          title={title !== 'no-border' ? title : undefined}
+          {...rest}
+        />
+      </ImageZoom>
     );
   },
   p: (props) => {
