+++ src/components/shared/image-zoom/index.js
@@ -1,0 +1,1 @@
+export { default } from './image-zoom';
