--- src/utils/image-block/index.js
@@ -1,1 +1,0 @@
-export { default } from './image-block';
